requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Data fetching
yfinance>=0.2.28

# Machine learning
tensorflow>=2.11.0
xgboost>=1.7.0

# Time series
statsmodels>=0.13.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Utilities
pyyaml>=6.0
python-dotenv>=0.19.0
joblib>=1.1.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development
jupyter>=1.0.0
black>=22.0.0
flake8>=4.0.0
